requirements.txt
PyYAML
rich
selenium
webdriver-manager
